kialo: refactor plotting to new file
