+++ kialo_domains_util.py
@@ -1,9 +1,6 @@
 import itertools
 from collections import defaultdict, Counter
 import pandas as pd
-import seaborn as sns
-import matplotlib.pyplot as plt
-import numpy as np
 
 
 def get_map2topics(path_kialo2topics):
@@ -80,28 +77,3 @@ def get_maps2uniquetopic(path_kialo2topics, maintopics):
     map2unique = get_unique_label(maps2maintopic)
     return map2unique, (maps2topics, sub2maintopic, main2subtopic)
 
-
-def plot_mapsPerTopic(path_kialo2topics, maintopics, outpath):
-    map2unique, _ = get_maps2uniquetopic(path_kialo2topics, maintopics)
-    d = Counter(map2unique.values())
-    d = dict(sorted(d.items(), key=lambda item: item[1]))
-    keys = list(d.keys())
-    # get values in the same order as keys, and parse percentage values
-    vals = list(d.values())
-    g = sns.barplot(x=keys, y=vals)
-    g.set_xticklabels(g.get_xticklabels(), rotation=30, fontsize=7, horizontalalignment='right')
-    plt.savefig(outpath, dpi=1000)
-
-
-
-
-if __name__ == '__main__':
-    kialo2topics = "/Users/falkne/PycharmProjects/deliberatorium/data/kilaoV2Langs/ENmapID2topicTags.txt"
-    maintopics = "/Users/falkne/PycharmProjects/deliberatorium/data/kilaoV2Langs/maintopic2subtopic.tsv"
-    outpath = "/Users/falkne/PycharmProjects/deliberatorium/data/kilaoV2Langs/topics/topicdistribution.png"
-    mapt2unique, (map2topics, _, _) = get_maps2uniquetopic(kialo2topics, maintopics)
-    # for map, tag in mapt2unique.items():
-    #    if tag == "politics":
-    #        othertags = map2topics[map]
-    #        print(othertags)
-    plot_mapsPerTopic(kialo2topics, maintopics, outpath)
